Refactor: Move deeplink logic into deeplink manager
