+++ Stellar/Manager/DeeplinkManager.swift
@@ -6,18 +6,54 @@
 //
 
 import Foundation
+import StellarMoonKit
 
-class DeepLinkManager: ObservableObject {
+final class DeepLinkManager: ObservableObject {
+
+	private var articlesVM = ArticleViewModel()
+	
 	@Published var selectedArticleID: String?
+	@Published var selectedRoute: AppRoute? = nil
 
-	func handleDeepLink(url: URL) {
-		if let components = NSURLComponents(url: url, resolvingAgainstBaseURL: true),
-		   components.scheme == "myapp",
-		   components.host == "article",
-		   let articleID = components.path?.replacingOccurrences(of: "/", with: "") {
-			selectedArticleID = articleID
+	func getArticleByDate(_ date: String) -> Article? {
+		let articles: [Article] = articlesVM.articles
+
+		return  articles.first { $0.date == date }
+	}
+
+	func handleUrl(_ url: URL) {
+		// Analyse de l'URL et de la définition de la route sélectionnée.
+		// Exemple: stellar://article?date=2023-04-01
+		guard let components = NSURLComponents(url: url, resolvingAgainstBaseURL: false),
+			  let host = components.host else {
+			return
+		}
+
+		switch host {
+		case "article":
+			if let id = components.queryItems?.first(where: { $0.name == "date" })?.value {
+				// Récupérer l'article correspondant à l'ID, puis définir la route sélectionnée.
+				if let article = getArticleByDate(id) {
+					selectedRoute = .articleDetail(article)
+				}
+			}
+		default:
+			break
+		}
+	}
+
+	func removeSelectedRoute() {
+		self.selectedRoute = nil
+	}
+
+	public enum AppRoute: Identifiable {
+		case articleDetail(Article)
+
+		var id: String {
+			switch self {
+			case .articleDetail(let article):
+				return article.title
+			}
 		}
 	}
 }
-
-// URL = stellar://article/23-02-2023
